pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "acp-agent-framework"
version = "0.1.0"
description = "Build custom ACP-compatible agents using your existing AI subscriptions"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Sanjay" }]
keywords = ["acp", "agent", "llm", "framework", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "agent-client-protocol>=0.8.0",
    "pydantic>=2.0",
    "click>=8.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "sse-starlette>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.4.0",
    "httpx>=0.24.0",
]
mcp = [
    "mcp>=1.0",
]
all = [
    "acp-agent-framework[dev,mcp]",
]

[project.scripts]
acp-agent = "acp_agent_framework.cli:main"

[project.urls]
Homepage = "https://github.com/sanjay3290/agentic-framework-acp"
Repository = "https://github.com/sanjay3290/agentic-framework-acp"
Issues = "https://github.com/sanjay3290/agentic-framework-acp/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/acp_agent_framework"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = ["integration: integration tests requiring real backends"]
addopts = "-m 'not integration'"

[tool.ruff]
target-version = "py310"
